viewファイル作成

+++ db/migrate/20201005014908_add_star_to_card.rb
@@ -1,5 +1,5 @@
 class AddStarToCard < ActiveRecord::Migration[6.0]
   def change
-    add_column :cards, :star, :integer
+    add_column :cards, :star, :float
   end
 end
